objc2_metal

Trait MTLDevice

pub unsafe trait MTLDevice:
    NSObjectProtocol
    + Send
    + Sync {

    // Provided methods
    fn name(&self) -> Retained<NSString>
       where Self: Sized + Message { ... }
    fn registryID(&self) -> u64
       where Self: Sized + Message { ... }
    fn architecture(&self) -> Retained<MTLArchitecture>
       where Self: Sized + Message { ... }
    fn isLowPower(&self) -> bool
       where Self: Sized + Message { ... }
    fn isHeadless(&self) -> bool
       where Self: Sized + Message { ... }
    fn isRemovable(&self) -> bool
       where Self: Sized + Message { ... }
    fn hasUnifiedMemory(&self) -> bool
       where Self: Sized + Message { ... }
    fn recommendedMaxWorkingSetSize(&self) -> u64
       where Self: Sized + Message { ... }
    fn location(&self) -> MTLDeviceLocation
       where Self: Sized + Message { ... }
    fn locationNumber(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn maxTransferRate(&self) -> u64
       where Self: Sized + Message { ... }
    fn isDepth24Stencil8PixelFormatSupported(&self) -> bool
       where Self: Sized + Message { ... }
    fn readWriteTextureSupport(&self) -> MTLReadWriteTextureTier
       where Self: Sized + Message { ... }
    fn argumentBuffersSupport(&self) -> MTLArgumentBuffersTier
       where Self: Sized + Message { ... }
    fn areRasterOrderGroupsSupported(&self) -> bool
       where Self: Sized + Message { ... }
    fn supports32BitFloatFiltering(&self) -> bool
       where Self: Sized + Message { ... }
    fn supports32BitMSAA(&self) -> bool
       where Self: Sized + Message { ... }
    fn supportsQueryTextureLOD(&self) -> bool
       where Self: Sized + Message { ... }
    fn supportsBCTextureCompression(&self) -> bool
       where Self: Sized + Message { ... }
    fn supportsPullModelInterpolation(&self) -> bool
       where Self: Sized + Message { ... }
    fn areBarycentricCoordsSupported(&self) -> bool
       where Self: Sized + Message { ... }
    fn supportsShaderBarycentricCoordinates(&self) -> bool
       where Self: Sized + Message { ... }
    fn currentAllocatedSize(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn heapTextureSizeAndAlignWithDescriptor(
        &self,
        desc: &MTLTextureDescriptor,
    ) -> MTLSizeAndAlign
       where Self: Sized + Message { ... }
    fn heapBufferSizeAndAlignWithLength_options(
        &self,
        length: NSUInteger,
        options: MTLResourceOptions,
    ) -> MTLSizeAndAlign
       where Self: Sized + Message { ... }
    fn newTextureWithDescriptor(
        &self,
        descriptor: &MTLTextureDescriptor,
    ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    fn newSharedTextureWithDescriptor(
        &self,
        descriptor: &MTLTextureDescriptor,
    ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    fn newSharedTextureWithHandle(
        &self,
        shared_handle: &MTLSharedTextureHandle,
    ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    fn newDefaultLibrary(
        &self,
    ) -> Option<Retained<ProtocolObject<dyn MTLLibrary>>>
       where Self: Sized + Message { ... }
    fn newDefaultLibraryWithBundle_error(
        &self,
        bundle: &NSBundle,
    ) -> Result<Retained<ProtocolObject<dyn MTLLibrary>>, Retained<NSError>>
       where Self: Sized + Message { ... }
    fn newLibraryWithFile_error(
        &self,
        filepath: &NSString,
    ) -> Result<Retained<ProtocolObject<dyn MTLLibrary>>, Retained<NSError>>
       where Self: Sized + Message { ... }
    fn newLibraryWithURL_error(
        &self,
        url: &NSURL,
    ) -> Result<Retained<ProtocolObject<dyn MTLLibrary>>, Retained<NSError>>
       where Self: Sized + Message { ... }
    fn newLibraryWithSource_options_error(
        &self,
        source: &NSString,
        options: Option<&MTLCompileOptions>,
    ) -> Result<Retained<ProtocolObject<dyn MTLLibrary>>, Retained<NSError>>
       where Self: Sized + Message { ... }
    fn supportsFeatureSet(&self, feature_set: MTLFeatureSet) -> bool
       where Self: Sized + Message { ... }
    fn supportsFamily(&self, gpu_family: MTLGPUFamily) -> bool
       where Self: Sized + Message { ... }
    fn supportsTextureSampleCount(&self, sample_count: NSUInteger) -> bool
       where Self: Sized + Message { ... }
    fn minimumLinearTextureAlignmentForPixelFormat(
        &self,
        format: MTLPixelFormat,
    ) -> NSUInteger
       where Self: Sized + Message { ... }
    fn minimumTextureBufferAlignmentForPixelFormat(
        &self,
        format: MTLPixelFormat,
    ) -> NSUInteger
       where Self: Sized + Message { ... }
    fn maxThreadgroupMemoryLength(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn maxArgumentBufferSamplerCount(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn areProgrammableSamplePositionsSupported(&self) -> bool
       where Self: Sized + Message { ... }
    unsafe fn supportsRasterizationRateMapWithLayerCount(
        &self,
        layer_count: NSUInteger,
    ) -> bool
       where Self: Sized + Message { ... }
    fn peerGroupID(&self) -> u64
       where Self: Sized + Message { ... }
    fn peerIndex(&self) -> u32
       where Self: Sized + Message { ... }
    fn peerCount(&self) -> u32
       where Self: Sized + Message { ... }
    fn sparseTileSizeInBytes(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn sparseTileSizeInBytesForSparsePageSize(
        &self,
        sparse_page_size: MTLSparsePageSize,
    ) -> NSUInteger
       where Self: Sized + Message { ... }
    fn maxBufferLength(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    unsafe fn sampleTimestamps_gpuTimestamp(
        &self,
        cpu_timestamp: NonNull<MTLTimestamp>,
        gpu_timestamp: NonNull<MTLTimestamp>,
    )
       where Self: Sized + Message { ... }
    fn supportsCounterSampling(
        &self,
        sampling_point: MTLCounterSamplingPoint,
    ) -> bool
       where Self: Sized + Message { ... }
    fn supportsVertexAmplificationCount(&self, count: NSUInteger) -> bool
       where Self: Sized + Message { ... }
    fn supportsDynamicLibraries(&self) -> bool
       where Self: Sized + Message { ... }
    fn supportsRenderDynamicLibraries(&self) -> bool
       where Self: Sized + Message { ... }
    fn supportsRaytracing(&self) -> bool
       where Self: Sized + Message { ... }
    unsafe fn heapAccelerationStructureSizeAndAlignWithSize(
        &self,
        size: NSUInteger,
    ) -> MTLSizeAndAlign
       where Self: Sized + Message { ... }
    fn supportsFunctionPointers(&self) -> bool
       where Self: Sized + Message { ... }
    fn supportsFunctionPointersFromRender(&self) -> bool
       where Self: Sized + Message { ... }
    fn supportsRaytracingFromRender(&self) -> bool
       where Self: Sized + Message { ... }
    fn supportsPrimitiveMotionBlur(&self) -> bool
       where Self: Sized + Message { ... }
    fn shouldMaximizeConcurrentCompilation(&self) -> bool
       where Self: Sized + Message { ... }
    fn setShouldMaximizeConcurrentCompilation(
        &self,
        should_maximize_concurrent_compilation: bool,
    )
       where Self: Sized + Message { ... }
    fn maximumConcurrentCompilationTaskCount(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn newResidencySetWithDescriptor_error(
        &self,
        desc: &MTLResidencySetDescriptor,
    ) -> Result<Retained<ProtocolObject<dyn MTLResidencySet>>, Retained<NSError>>
       where Self: Sized + Message { ... }
    fn queryTimestampFrequency(&self) -> u64
       where Self: Sized + Message { ... }

}

MTLDevice represents a processor capable of data parallel computations

See also Apples documentation

Provided Methods

fn name(&self) -> Retained<NSString>
where Self: Sized + Message,

The full name of the vendor device.

fn registryID(&self) -> u64
where Self: Sized + Message,

Returns the IORegistry ID for the Metal device

The registryID value for a Metal device is global to all tasks, and may be used to identify the GPU across task boundaries.

fn architecture(&self) -> Retained<MTLArchitecture>
where Self: Sized + Message,

Returns the devices architecture information.

fn isLowPower(&self) -> bool
where Self: Sized + Message,

On systems that support automatic graphics switching, this will return YES for the the low power device.

fn isHeadless(&self) -> bool
where Self: Sized + Message,

On systems that include more that one GPU, this will return YES for any device that does not support any displays. Only available on Mac OS X.

fn isRemovable(&self) -> bool
where Self: Sized + Message,

If this GPU is removable, this property will return YES.

If a GPU is is removed without warning, APIs may fail even with good input, even before a notification can get posted informing the application that the device has been removed.

fn hasUnifiedMemory(&self) -> bool
where Self: Sized + Message,

Returns YES if this GPU shares its memory with the rest of the machine (CPU, etc.)

Some GPU architectures do not have dedicated local memory and instead only use the same memory shared with the rest of the machine. This property will return YES for GPUs that fall into that category.

fn recommendedMaxWorkingSetSize(&self) -> u64
where Self: Sized + Message,

Returns an approximation of how much memory this device can use with good performance.

Performance may be improved by keeping the total size of all resources (texture and buffers) and heaps less than this threshold, beyond which the device is likely to be overcommitted and incur a performance penalty.

fn location(&self) -> MTLDeviceLocation
where Self: Sized + Message,

Returns an enum that indicates where the GPU is located relative to the host computer.

The returned value indicates if the GPU is built into the computer, inserted into a slot internal to the computer, or external to the computer. Otherwise it is Unspecified

fn locationNumber(&self) -> NSUInteger
where Self: Sized + Message,

Returns a value that further specifies the GPUs location

The returned value indicates which slot or Thunderbolt port the GPU is attached to. For Built-in GPUs, if LowPower this value is 0, otherwise it is 1. It is possible for multiple GPUs to have the same location and locationNumber; e.g.: A PCI card with multiple GPUs, or an eGPU daisy-chained off of another eGPU attached to a host Thunderbolt port.

fn maxTransferRate(&self) -> u64
where Self: Sized + Message,

Upper bound of System RAM < => VRAM transfer rate in bytes/sec

The returned value indicates the theoretical maximum data rate in bytes/second from host memory to the GPUs VRAM. This is derived from the raw data clock rate and as such may not be reachable under real-world conditions. For Built-in GPUs this value is 0.

fn isDepth24Stencil8PixelFormatSupported(&self) -> bool
where Self: Sized + Message,

If YES, device supports MTLPixelFormatDepth24Unorm_Stencil8.

fn readWriteTextureSupport(&self) -> MTLReadWriteTextureTier
where Self: Sized + Message,

Query support tier for read-write texture formats.

Returns: MTLReadWriteTextureTier enum value.

fn argumentBuffersSupport(&self) -> MTLArgumentBuffersTier
where Self: Sized + Message,

Query support tier for Argument Buffers.

Returns: MTLArgumentBuffersTier enum value.

fn areRasterOrderGroupsSupported(&self) -> bool
where Self: Sized + Message,

Query device for raster order groups support.

Returns: BOOL value. If YES, the device supports raster order groups. If NO, the device does not.

fn supports32BitFloatFiltering(&self) -> bool
where Self: Sized + Message,

Query device for 32-bit Float texture filtering support. Specifically, R32Float, RG32Float, and RGBA32Float.

Returns: BOOL value. If YES, the device supports filtering 32-bit Float textures. If NO, the device does not.

fn supports32BitMSAA(&self) -> bool
where Self: Sized + Message,

Query device for 32-bit MSAA texture support. Specifically, added support for allocating 32-bit Integer format textures (R32Uint, R32Sint, RG32Uint, RG32Sint, RGBA32Uint, and RGBA32Sint) and resolving 32-bit Float format textures (R32Float, RG32Float, and RGBA32Float).

Returns: BOOL value. If YES, the device supports these additional 32-bit MSAA texture capabilities. If NO, the devices does not.

fn supportsQueryTextureLOD(&self) -> bool
where Self: Sized + Message,

Query device for whether it supports the calculate_clampled_lod and calculate_unclamped_lod Metal shading language functionality.

Returns: BOOL value. If YES, the device supports the calculate LOD functionality. If NO, the device does not.

fn supportsBCTextureCompression(&self) -> bool
where Self: Sized + Message,

Query device for BC Texture format support

Returns: BOOL value. If YES, the device supports compressed BC Texture formats. If NO, the device does not.

fn supportsPullModelInterpolation(&self) -> bool
where Self: Sized + Message,

Query device for pull model interpolation support which allows a fragment shader to compute multiple interpolations (at center, at centroid, at offset, at sample) of a fragment input.

Returns: BOOL value. If YES, the device supports pull model interpolation. If NO, the device does not.

fn areBarycentricCoordsSupported(&self) -> bool
where Self: Sized + Message,

Deprecated

Query device for Barycentric coordinates support; deprecated, use supportsShaderBarycentricCoordinates

Returns: BOOL value. If YES, the device barycentric coordinates

fn supportsShaderBarycentricCoordinates(&self) -> bool
where Self: Sized + Message,

Query device for Barycentric Coordinates support.

Returns: BOOL value. If YES, the device supports barycentric coordinates. If NO, the device does not.

fn currentAllocatedSize(&self) -> NSUInteger
where Self: Sized + Message,

The current size in bytes of all resources allocated by this device

fn heapTextureSizeAndAlignWithDescriptor( &self, desc: &MTLTextureDescriptor, ) -> MTLSizeAndAlign
where Self: Sized + Message,

Determine the byte size of textures when sub-allocated from a heap.

This method can be used to help determine the required heap size.

fn heapBufferSizeAndAlignWithLength_options( &self, length: NSUInteger, options: MTLResourceOptions, ) -> MTLSizeAndAlign
where Self: Sized + Message,

Determine the byte size of buffers when sub-allocated from a heap.

This method can be used to help determine the required heap size.

fn newTextureWithDescriptor( &self, descriptor: &MTLTextureDescriptor, ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

Allocate a new texture with privately owned storage.

fn newSharedTextureWithDescriptor( &self, descriptor: &MTLTextureDescriptor, ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

Create a new texture that can be shared across process boundaries.

This texture can be shared between process boundaries but not between different GPUs, by passing its MTLSharedTextureHandle.

Parameter descriptor: A description of the properties for the new texture.

Returns: A new texture object.

fn newSharedTextureWithHandle( &self, shared_handle: &MTLSharedTextureHandle, ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

Recreate shared texture from received texture handle.

This texture was shared between process boundaries by other process using MTLSharedTextureHandle. Current process will now share it with other processes and will be able to interact with it (but still in scope of the same GPUs).

Parameter sharedHandle: Handle to shared texture in this process space.

Returns: A new texture object.

fn newDefaultLibrary(&self) -> Option<Retained<ProtocolObject<dyn MTLLibrary>>>
where Self: Sized + Message,

Returns the default library for the main bundle.

use newDefaultLibraryWithBundle:error: to get an NSError in case of failure.

fn newDefaultLibraryWithBundle_error( &self, bundle: &NSBundle, ) -> Result<Retained<ProtocolObject<dyn MTLLibrary>>, Retained<NSError>>
where Self: Sized + Message,

fn newLibraryWithFile_error( &self, filepath: &NSString, ) -> Result<Retained<ProtocolObject<dyn MTLLibrary>>, Retained<NSError>>
where Self: Sized + Message,

Deprecated: Use -newLibraryWithURL:error: instead

Load a MTLLibrary from a metallib file.

fn newLibraryWithURL_error( &self, url: &NSURL, ) -> Result<Retained<ProtocolObject<dyn MTLLibrary>>, Retained<NSError>>
where Self: Sized + Message,

Load a MTLLibrary from a metallib file.

fn newLibraryWithSource_options_error( &self, source: &NSString, options: Option<&MTLCompileOptions>, ) -> Result<Retained<ProtocolObject<dyn MTLLibrary>>, Retained<NSError>>
where Self: Sized + Message,

Load a MTLLibrary from source.

fn supportsFeatureSet(&self, feature_set: MTLFeatureSet) -> bool
where Self: Sized + Message,

Deprecated: Use supportsFamily instead

Returns TRUE if the feature set is supported by this MTLDevice.

fn supportsFamily(&self, gpu_family: MTLGPUFamily) -> bool
where Self: Sized + Message,

Returns TRUE if the GPU Family is supported by this MTLDevice.

fn supportsTextureSampleCount(&self, sample_count: NSUInteger) -> bool
where Self: Sized + Message,

Query device if it support textures with a given sampleCount.

Returns: BOOL value. If YES, device supports the given sampleCount for textures. If NO, device does not support the given sampleCount.

fn minimumLinearTextureAlignmentForPixelFormat( &self, format: MTLPixelFormat, ) -> NSUInteger
where Self: Sized + Message,

Returns the minimum alignment required for offset and rowBytes when creating a linear texture. An error is thrown for queries with invalid pixel formats (depth, stencil, or compressed formats).

fn minimumTextureBufferAlignmentForPixelFormat( &self, format: MTLPixelFormat, ) -> NSUInteger
where Self: Sized + Message,

Returns the minimum alignment required for offset and rowBytes when creating a texture buffer from a buffer.

fn maxThreadgroupMemoryLength(&self) -> NSUInteger
where Self: Sized + Message,

The maximum threadgroup memory available, in bytes.

fn maxArgumentBufferSamplerCount(&self) -> NSUInteger
where Self: Sized + Message,

The maximum number of unique argument buffer samplers per app.

This limit is only applicable to samplers that have their supportArgumentBuffers property set to true. A MTLSamplerState object is considered unique if the configuration of its originating MTLSamplerDescriptor properties is unique. For example, two samplers with equal minFilter values but different magFilter values are considered unique.

fn areProgrammableSamplePositionsSupported(&self) -> bool
where Self: Sized + Message,

Query device for programmable sample position support.

Returns: BOOL value. If YES, the device supports programmable sample positions. If NO, the device does not.

unsafe fn supportsRasterizationRateMapWithLayerCount( &self, layer_count: NSUInteger, ) -> bool
where Self: Sized + Message,

Query device for variable rasterization rate support with the given number of layers.

Parameter layerCount: The number of layers for which to query device support.

Returns: YES if the device supports creation of rendering using a MTLRasterizationRateMap with the given number of layers.

Safety

layerCount might not be bounds-checked.

fn peerGroupID(&self) -> u64
where Self: Sized + Message,

If a device supports peer to peer transfers with another device (or devices), this property will return a unique 64-bit identifier associated with all devices in the same peer group.

fn peerIndex(&self) -> u32
where Self: Sized + Message,

All Metal devices that are part of the same peer group will have a unique index value within the group in the range from 0 to peerCount - 1.

fn peerCount(&self) -> u32
where Self: Sized + Message,

For Metal devices that are part of a peer group, this property returns the total number of devices in that group.

fn sparseTileSizeInBytes(&self) -> NSUInteger
where Self: Sized + Message,

Returns the number of bytes required to map one sparse texture tile.

fn sparseTileSizeInBytesForSparsePageSize( &self, sparse_page_size: MTLSparsePageSize, ) -> NSUInteger
where Self: Sized + Message,

Returns the number of bytes required to map one sparse texture tile for a given MTLSparsePageSize

fn maxBufferLength(&self) -> NSUInteger
where Self: Sized + Message,

unsafe fn sampleTimestamps_gpuTimestamp( &self, cpu_timestamp: NonNull<MTLTimestamp>, gpu_timestamp: NonNull<MTLTimestamp>, )
where Self: Sized + Message,

Sample the CPU and GPU timestamps as closely as possible.

Parameter cpuTimestamp: The timestamp on the CPU

Parameter gpuTimestamp: The timestamp on the GPU

Safety

cpu_timestamp must be a valid pointer.

gpu_timestamp must be a valid pointer.

fn supportsCounterSampling( &self, sampling_point: MTLCounterSamplingPoint, ) -> bool
where Self: Sized + Message,

Query device for counter sampling points support.

Parameter samplingPoint: Query index

Returns: BOOL value. If YES, the device supports counter sampling at given point.

fn supportsVertexAmplificationCount(&self, count: NSUInteger) -> bool
where Self: Sized + Message,

Query device for vertex amplification support.

Parameter count: The amplification count to check

Returns: BOOL value. If YES, the device supports vertex amplification with the given count. If NO, the device does not.

fn supportsDynamicLibraries(&self) -> bool
where Self: Sized + Message,

Query device support for creating and using dynamic libraries in a compute pipeline.

Returns: BOOL value. If YES, the device supports creating and using dynamic libraries in a compute pipeline. If NO, the device does not.

fn supportsRenderDynamicLibraries(&self) -> bool
where Self: Sized + Message,

Query device support for creating and using dynamic libraries in render pipeline stages.

Returns: BOOL value. If YES, the device supports creating and using dynamic libraries in render pipeline stages. If NO, the device does not.

fn supportsRaytracing(&self) -> bool
where Self: Sized + Message,

Query device support for using ray tracing from compute pipelines.

Returns: BOOL value. If YES, the device supports ray tracing from compute pipelines. If NO, the device does not.

unsafe fn heapAccelerationStructureSizeAndAlignWithSize( &self, size: NSUInteger, ) -> MTLSizeAndAlign
where Self: Sized + Message,

Determine the byte size of acceleration structures when sub-allocated from a heap.

This method can be used to help determine the required heap size.

Safety

size might not be bounds-checked.

fn supportsFunctionPointers(&self) -> bool
where Self: Sized + Message,

Query device support for using function pointers from compute pipelines.

Returns: BOOL value. If YES, the device supports function pointers from compute pipelines. If NO, the device does not.

fn supportsFunctionPointersFromRender(&self) -> bool
where Self: Sized + Message,

Query device support for using function pointers from render pipeline stages.

Returns: BOOL value. If YES, the device supports function pointers from render pipeline stages. If NO, the device does not.

fn supportsRaytracingFromRender(&self) -> bool
where Self: Sized + Message,

Query device support for using ray tracing from render pipeline stages.

Returns: BOOL value. If YES, the device supports ray tracing from render pipeline stages. If NO, the device does not.

fn supportsPrimitiveMotionBlur(&self) -> bool
where Self: Sized + Message,

Query device support for using ray tracing primitive motion blur.

Returns: BOOL value. If YES, the device supports the primitive motion blur api. If NO, the device does not.

fn shouldMaximizeConcurrentCompilation(&self) -> bool
where Self: Sized + Message,

Allow this device to use additional CPU threads (scaled automatically to the host machine) to be used for compilation tasks. Default is NO.

Use the maximumConcurrentCompilationTaskCount property to determine the current number of concurrent CPU threads that this device is using.

fn setShouldMaximizeConcurrentCompilation( &self, should_maximize_concurrent_compilation: bool, )
where Self: Sized + Message,

Setter for shouldMaximizeConcurrentCompilation.

fn maximumConcurrentCompilationTaskCount(&self) -> NSUInteger
where Self: Sized + Message,

Returns the maximum count of concurrent executing compilation tasks.

The property returns a different value depending on the value of the property shouldMaximizeConcurrentCompilation.

fn newResidencySetWithDescriptor_error( &self, desc: &MTLResidencySetDescriptor, ) -> Result<Retained<ProtocolObject<dyn MTLResidencySet>>, Retained<NSError>>
where Self: Sized + Message,

Creates a new residency set with a descriptor.

fn queryTimestampFrequency(&self) -> u64
where Self: Sized + Message,

Queries the frequency of the GPU timestamp in ticks per second.

Returns: The frequency of the GPU timestamp in ticks per second.