objc2_metal

Trait MTLFunction

pub unsafe trait MTLFunction:
    NSObjectProtocol
    + Send
    + Sync {
    // Provided methods
    fn label(&self) -> Option<Retained<NSString>>
       where Self: Sized + Message { ... }
    fn setLabel(&self, label: Option<&NSString>)
       where Self: Sized + Message { ... }
    fn device(&self) -> Retained<ProtocolObject<dyn MTLDevice>>
       where Self: Sized + Message { ... }
    fn functionType(&self) -> MTLFunctionType
       where Self: Sized + Message { ... }
    fn patchType(&self) -> MTLPatchType
       where Self: Sized + Message { ... }
    fn patchControlPointCount(&self) -> NSInteger
       where Self: Sized + Message { ... }
    fn vertexAttributes(&self) -> Option<Retained<NSArray<MTLVertexAttribute>>>
       where Self: Sized + Message { ... }
    fn stageInputAttributes(&self) -> Option<Retained<NSArray<MTLAttribute>>>
       where Self: Sized + Message { ... }
    fn name(&self) -> Retained<NSString>
       where Self: Sized + Message { ... }
    fn functionConstantsDictionary(
        &self,
    ) -> Retained<NSDictionary<NSString, MTLFunctionConstant>>
       where Self: Sized + Message { ... }
}

A handle to intermediate code used as inputs for either a MTLComputePipelineState or a MTLRenderPipelineState.

MTLFunction is a single vertex shader, fragment shader, or compute function. A Function can only be used with the device that it was created against.

See also Apple’s documentation

Provided Methods

fn label(&self) -> Option<Retained<NSString>>
where Self: Sized + Message,

A string to help identify this object.

fn setLabel(&self, label: Option<&NSString>)
where Self: Sized + Message,

Setter for label.

This is copied when set.

fn device(&self) -> Retained<ProtocolObject<dyn MTLDevice>>
where Self: Sized + Message,

The device this resource was created against. This resource can only be used with this device.

fn functionType(&self) -> MTLFunctionType
where Self: Sized + Message,

The overall kind of entry point: compute, vertex, or fragment.

fn patchType(&self) -> MTLPatchType
where Self: Sized + Message,

Returns the patch type. MTLPatchTypeNone if it is not a post tessellation vertex shader.

fn patchControlPointCount(&self) -> NSInteger
where Self: Sized + Message,

Returns the number of patch control points if it was specified in the shader. Returns -1 if it was not specified.

fn vertexAttributes(&self) -> Option<Retained<NSArray<MTLVertexAttribute>>>
where Self: Sized + Message,

fn stageInputAttributes(&self) -> Option<Retained<NSArray<MTLAttribute>>>
where Self: Sized + Message,

Returns an array describing the attributes

fn name(&self) -> Retained<NSString>
where Self: Sized + Message,

The name of the function in the shading language.

fn functionConstantsDictionary( &self, ) -> Retained<NSDictionary<NSString, MTLFunctionConstant>>
where Self: Sized + Message,

A dictionary containing information about all function contents, keyed by the constant names.