objc2_metal

Trait MTLTexture

pub unsafe trait MTLTexture: MTLResource {

    // Provided methods
    fn rootResource(&self) -> Option<Retained<ProtocolObject<dyn MTLResource>>>
       where Self: Sized + Message { ... }
    fn parentTexture(&self) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    fn parentRelativeLevel(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn parentRelativeSlice(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn bufferOffset(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn bufferBytesPerRow(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn iosurfacePlane(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn textureType(&self) -> MTLTextureType
       where Self: Sized + Message { ... }
    fn pixelFormat(&self) -> MTLPixelFormat
       where Self: Sized + Message { ... }
    fn width(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn height(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn depth(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn mipmapLevelCount(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn sampleCount(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn arrayLength(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn usage(&self) -> MTLTextureUsage
       where Self: Sized + Message { ... }
    fn isShareable(&self) -> bool
       where Self: Sized + Message { ... }
    fn isFramebufferOnly(&self) -> bool
       where Self: Sized + Message { ... }
    fn firstMipmapInTail(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn tailSizeInBytes(&self) -> NSUInteger
       where Self: Sized + Message { ... }
    fn isSparse(&self) -> bool
       where Self: Sized + Message { ... }
    fn allowGPUOptimizedContents(&self) -> bool
       where Self: Sized + Message { ... }
    fn compressionType(&self) -> MTLTextureCompressionType
       where Self: Sized + Message { ... }
    fn newTextureViewWithPixelFormat(
        &self,
        pixel_format: MTLPixelFormat,
    ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    unsafe fn newTextureViewWithPixelFormat_textureType_levels_slices(
        &self,
        pixel_format: MTLPixelFormat,
        texture_type: MTLTextureType,
        level_range: NSRange,
        slice_range: NSRange,
    ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    fn newSharedTextureHandle(&self) -> Option<Retained<MTLSharedTextureHandle>>
       where Self: Sized + Message { ... }
    fn newTextureViewWithDescriptor(
        &self,
        descriptor: &MTLTextureViewDescriptor,
    ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    fn remoteStorageTexture(
        &self,
    ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    fn newRemoteTextureViewForDevice(
        &self,
        device: &ProtocolObject<dyn MTLDevice>,
    ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    fn swizzle(&self) -> MTLTextureSwizzleChannels
       where Self: Sized + Message { ... }
    unsafe fn newTextureViewWithPixelFormat_textureType_levels_slices_swizzle(
        &self,
        pixel_format: MTLPixelFormat,
        texture_type: MTLTextureType,
        level_range: NSRange,
        slice_range: NSRange,
        swizzle: MTLTextureSwizzleChannels,
    ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
       where Self: Sized + Message { ... }
    fn sparseTextureTier(&self) -> MTLTextureSparseTier
       where Self: Sized + Message { ... }

}

MTLTexture represents a collection of 1D, 2D, or 3D images.

Each image in a texture is a 1D, 2D, 2DMultisample, or 3D image. The texture contains one or more images arranged in a mipmap stack. If there are multiple mipmap stacks, each one is referred to as a slice of the texture. 1D, 2D, 2DMultisample, and 3D textures have a single slice. In 1DArray and 2DArray textures, every slice is an array element. A Cube texture always has 6 slices, one for each face. In a CubeArray texture, each set of six slices is one element in the array.

Most APIs that operate on individual images in a texture address those images via a tuple of a Slice, and Mipmap Level within that slice.

See also Apple’s documentation

Provided Methods

fn rootResource(&self) -> Option<Retained<ProtocolObject<dyn MTLResource>>>
where Self: Sized + Message,

👎Deprecated: Use parentTexture or buffer instead

The resource this texture was created from. It may be a texture or a buffer. If this texture is not reusing storage of another MTLResource, then nil is returned.

fn parentTexture(&self) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

The texture this texture view was created from, or nil if this is not a texture view or it was not created from a texture.

fn parentRelativeLevel(&self) -> NSUInteger
where Self: Sized + Message,

The base level of the texture this texture view was created from, or 0 if this is not a texture view.

fn parentRelativeSlice(&self) -> NSUInteger
where Self: Sized + Message,

The base slice of the texture this texture view was created from, or 0 if this is not a texture view.

fn bufferOffset(&self) -> NSUInteger
where Self: Sized + Message,

The offset of the buffer this texture view was created from, or 0 if this is not a texture view.

fn bufferBytesPerRow(&self) -> NSUInteger
where Self: Sized + Message,

The bytesPerRow of the buffer this texture view was created from, or 0 if this is not a texture view.

fn iosurfacePlane(&self) -> NSUInteger
where Self: Sized + Message,

If this texture was created from an IOSurface, this returns the plane of the IOSurface from which the texture was created. iosurfacePlane is 0 if this texture was not created from an IOSurface.

fn textureType(&self) -> MTLTextureType
where Self: Sized + Message,

The type of this texture.

fn pixelFormat(&self) -> MTLPixelFormat
where Self: Sized + Message,

The MTLPixelFormat that is used to interpret this texture’s contents.

fn width(&self) -> NSUInteger
where Self: Sized + Message,

The width of the MTLTexture instance in pixels.

fn height(&self) -> NSUInteger
where Self: Sized + Message,

The height of the MTLTexture instance in pixels.

. height is 1 if the texture is 1D.

fn depth(&self) -> NSUInteger
where Self: Sized + Message,

The depth of this MTLTexture instance in pixels.

If this MTLTexture is not a 3D texture, the depth is 1

fn mipmapLevelCount(&self) -> NSUInteger
where Self: Sized + Message,

The number of mipmap levels in each slice of this MTLTexture.

fn sampleCount(&self) -> NSUInteger
where Self: Sized + Message,

The number of samples in each pixel of this MTLTexture.

If this texture is any type other than 2DMultisample, samples is 1.

fn arrayLength(&self) -> NSUInteger
where Self: Sized + Message,

The number of array elements in this MTLTexture.

For non-Array texture types, arrayLength is 1.

fn usage(&self) -> MTLTextureUsage
where Self: Sized + Message,

Description of texture usage.

fn isShareable(&self) -> bool
where Self: Sized + Message,

If YES, this texture can be shared with other processes.

Texture can be shared across process addres space boundaries through use of sharedTextureHandle and XPC.

fn isFramebufferOnly(&self) -> bool
where Self: Sized + Message,

If YES, this texture can only be used with a MTLAttachmentDescriptor, and cannot be used as a texture argument for MTLRenderCommandEncoder, MTLBlitCommandEncoder, or MTLComputeCommandEncoder. Furthermore, when this property’s value is YES, readPixels/writePixels may not be used with this texture.

Textures obtained from CAMetalDrawables may have this property set to YES, depending on the value of frameBufferOnly passed to their parent CAMetalLayer. Textures created directly by the application will not have any restrictions.

fn firstMipmapInTail(&self) -> NSUInteger
where Self: Sized + Message,

For sparse textures this property returns index of first mipmap that is packed in tail. Mapping this mipmap level will map all subsequent mipmap levels.

fn tailSizeInBytes(&self) -> NSUInteger
where Self: Sized + Message,

Amount of memory in bytes required to map sparse texture tail.

fn isSparse(&self) -> bool
where Self: Sized + Message,

fn allowGPUOptimizedContents(&self) -> bool
where Self: Sized + Message,

Allow GPU-optimization for the contents texture. The default value is true.

Useful for opting-out of GPU-optimization when implicit optimization (e.g. RT writes) is regressing CPU-read-back performance. See the documentation for optimizeContentsForGPUAccess: and optimizeContentsForCPUAccess: APIs.

fn compressionType(&self) -> MTLTextureCompressionType
where Self: Sized + Message,

Returns the compression type of the texture

See the compressionType property on MTLTextureDescriptor

fn newTextureViewWithPixelFormat( &self, pixel_format: MTLPixelFormat, ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

Create a new texture which shares the same storage as the source texture, but with a different (but compatible) pixel format.

unsafe fn newTextureViewWithPixelFormat_textureType_levels_slices( &self, pixel_format: MTLPixelFormat, texture_type: MTLTextureType, level_range: NSRange, slice_range: NSRange, ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

Create a new texture which shares the same storage as the source texture, but with a different (but compatible) pixel format, texture type, levels and slices.

Safety

levelRange might not be bounds-checked.

sliceRange might not be bounds-checked.

fn newSharedTextureHandle(&self) -> Option<Retained<MTLSharedTextureHandle>>
where Self: Sized + Message,

Create a new texture handle, that can be shared across process addres space boundaries.

fn newTextureViewWithDescriptor( &self, descriptor: &MTLTextureViewDescriptor, ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

Create a new texture which shares the same storage as the source texture, but with different (but compatible) properties specified by the descriptor

fn remoteStorageTexture( &self, ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

For Metal texture objects that are remote views, this returns the texture associated with the storage on the originating device.

fn newRemoteTextureViewForDevice( &self, device: &ProtocolObject<dyn MTLDevice>, ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

On Metal devices that support peer to peer transfers, this method is used to create a remote texture view on another device within the peer group. The receiver must use MTLStorageModePrivate or be backed by an IOSurface.

fn swizzle(&self) -> MTLTextureSwizzleChannels
where Self: Sized + Message,

The channel swizzle used when reading or sampling from this texture

unsafe fn newTextureViewWithPixelFormat_textureType_levels_slices_swizzle( &self, pixel_format: MTLPixelFormat, texture_type: MTLTextureType, level_range: NSRange, slice_range: NSRange, swizzle: MTLTextureSwizzleChannels, ) -> Option<Retained<ProtocolObject<dyn MTLTexture>>>
where Self: Sized + Message,

Create a new texture which shares the same storage as the source texture, but with a different (but compatible) pixel format, texture type, levels, slices and swizzle.

Safety

levelRange might not be bounds-checked.

sliceRange might not be bounds-checked.

fn sparseTextureTier(&self) -> MTLTextureSparseTier
where Self: Sized + Message,

Query support tier for sparse textures.