Crate objc2_metal

Bindings to the Metal framework

See Apple’s docs and the general docs on framework crates for more information.

Metal has tools for validating that you’re using it correctly, using these is highly recommended! See Apple’s documentation on it, or run man MetalValidation to get information on environment variables.

NOTE: To use MTLCreateSystemDefaultDevice you need to link to CoreGraphics, this can be done by using objc2-core-graphics, or by doing:

#[link(name = "CoreGraphics", kind = "framework")]
extern "C" {}

Safety considerations

Metal allows running arbitrary code on the GPU. We treat memory safety issues on the GPU as just as unsafe as that which applies to the CPU. A few notes on this below.

Shaders

Shaders are (often) written in an unsafe C-like language.

Loading them (via MTLLibrary, function stitching etc.) is perfectly safe, it is similar to dynamic linking. The restrictions that e.g. libloading::Library::new labours under do not apply, since there are no ctors in the Metal Shading Language (see section 4.2).

Similarly, getting individual shaders (MTLFunction) is safe, we can model this as the same as calling dlsym (which just returns a pointer).

Calling functions though, is not safe. Even though they can have their parameter and return types checked at runtime, they may have additional restrictions not present in the signature (e.g. __builtin_unreachable() is possible in MSL, so is out-of-bounds accesses). If you view MTLFunction as essentially just an unsafe fn() pointer, this should be apparent.

Bounds checks

It is yet unclear whether Metal APIs are bounds-checked on the CPU side or not, so APIs that take offsets / lengths are often unsafe.

Synchronization

MTLResource subclasses such as MTLBuffer and MTLTexture require synchronization between the CPU and the GPU, or between different threads on the GPU itself, so APIs taking these are often unsafe.

Memory management and lifetimes

Resources used in MTL4CommandBuffers or command buffers with created with one of:

MTLCommandBufferDescriptor::setRetainedReferences(false).

MTLCommandQueue::commandBufferWithUnretainedReferences().

Must be kept alive for as long as they’re used.

Type safety

MTLBuffer is untyped (in a similar manner as a [u8] slice), you must ensure that any usage of it is done with valid types.

Structs

MTLAccelerationStructureSizes

Describes the memory requirements for an acceleration structure

MTLArchitecture

Contains information about the device’s architecture

MTLArgumentBuffersTier

MTLArgumentBuffersTier determines support level for argument buffers.

MTLArgumentDescriptor

Represents a member of an argument buffer

MTLAttribute

Apple’s documentation

MTLBufferSparseTier

Enumerates the different support levels for sparse buffers.

MTLCPUCacheMode

Describes what CPU cache mode is used for the CPU’s mapping of a texture resource.

MTLCompileOptions

Apple’s documentation

MTLCompileSymbolVisibility

Apple’s documentation

MTLCounterSamplingPoint

MTLCounterSamplingPoint determines type of sampling points that are supported on given device.

MTLDeviceLocation

Specifies the location of the GPU on macOS

MTLFeatureSetDeprecated

Apple’s documentation

MTLFunctionConstant

describe an uberShader constant used by the function

MTLFunctionReflection

Represents a reflection object containing information about a function in a Metal library.

MTLFunctionType

An identifier for a top-level Metal function.

MTLGPUFamily

Apple’s documentation

MTLHazardTrackingMode

Describes how hazard tracking is performed.

MTLIOCompressionMethod

Apple’s documentation

MTLLanguageVersion

Apple’s documentation

MTLLibraryError

NSErrors raised when creating a library.

MTLLibraryOptimizationLevel

Optimization level for the Metal compiler.

MTLLibraryType

Apple’s documentation

MTLMathFloatingPointFunctions

An enum to indicate the default math functions for single precision floating-point

MTLMathMode

An enum to indicate if the compiler can perform optimizations for floating-point arithmetic that may violate the IEEE 754 standard

MTLPatchType

Apple’s documentation

MTLPipelineOption

Controls the creation of the pipeline

MTLPixelFormat

Apple’s documentation

MTLPurgeableState

Options for setPurgeable call.

MTLReadWriteTextureTier

MTLReadWriteTextureTier determines support level for read-write texture formats.

MTLResidencySetDescriptor

Specifies the parameters for MTLResidencySet creation.

MTLResourceOptions

Apple’s documentation

MTLSharedTextureHandle

Apple’s documentation

MTLSizeAndAlign

Represent a memory size and alignment in bytes.

MTLSparsePageSize

Physical size of sparse resource page in KBs.

MTLSparseTextureRegionAlignmentMode

MTLSparseTextureRegionAlignmentMode determines type of alignment used when converting from pixel region to tile region.

MTLStorageMode

Describes location and CPU mapping of MTLTexture.

MTLTextureCompressionType

Apple’s documentation

MTLTextureDescriptor

Apple’s documentation

MTLTextureSparseTier

Enumerates the different support levels for sparse textures.

MTLTextureSwizzle

Apple’s documentation

MTLTextureSwizzleChannels

Apple’s documentation

MTLTextureType

MTLTextureType describes the dimensionality of each image, and if multiple images are arranged into an array or cube.

MTLTextureUsage

MTLTextureUsage declares how the texture will be used over its lifetime (bitwise OR for multiple uses).

MTLTextureViewDescriptor

Apple’s documentation

MTLVertexAttribute

Apple’s documentation

Constants

MTLResourceCPUCacheModeMask

MTLResourceCPUCacheModeShift

MTLResourceHazardTrackingModeMask

MTLResourceHazardTrackingModeShift

MTLResourceStorageModeMask

MTLResourceStorageModeShift

MTLTextureSwizzleChannelsDefault

Statics

MTLDeviceRemovalRequestedNotification⚠

This notification is posted when the user has requested that applications cease using a particular device. Applications should assume that the device will be removed (terminated) imminently. Additionally, the device will be removed from the internal device array prior to this notification being posted. Applications should immediately begin the process of releasing all resources created on the given device, as well as any references to the device itself.

MTLDeviceWasAddedNotification⚠

This notification is posted when a new Metal device is added to the system

MTLDeviceWasRemovedNotification⚠

This notification is posted if the device is removed while there are still outstanding references to it, due to either a surprise or forced disconnect by the user. Applications must expect that any attempt to use the device after this point will fail.

MTLLibraryErrorDomain⚠

NSErrors raised when creating a library.

Traits

MTLAllocation

Base class for Metal allocations.

MTLDevice

MTLDevice represents a processor capable of data parallel computations

MTLDrawable

All “drawable” objects (such as those coming from CAMetalLayer) are expected to conform to this protocol

MTLFunction

A handle to intermediate code used as inputs for either a MTLComputePipelineState or a MTLRenderPipelineState.

MTLLibrary

Apple’s documentation

MTLResidencySet

A residency set is responsible for managing resource and heap residency and is referenced by a command buffer or command queue in order to ensure that resources and heaps are resident. Resources and heaps are added and removed uncommitted and a subsequent commit call applies all of the changes in bulk.

MTLResource

Common APIs available for MTLBuffer and MTLTexture instances

MTLTexture

MTLTexture represents a collection of 1D, 2D, or 3D images.

Functions

MTLCopyAllDevices

Returns all Metal devices in the system.

MTLCreateSystemDefaultDevice

Returns a reference to the preferred system default Metal device.

MTLRemoveDeviceObserver⚠

Removes a previously installed observer for device change notifications.

Type Aliases

MTLDeviceNotificationName

Type for device notifications

MTLGPUAddress

A 64-bit unsigned integer type appropriate for storing GPU addresses.

MTLTimestamp

Apple’s documentation